rest_api/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
Faker==19.10.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
